requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pandas
tqdm
youtube_dl
webvtt-py
argparse